pyproject.toml
[project]
name = "steam_screenshot_manager"
version = "0.1.0"
authors = [
	{ name = "Celyn Walters", email = "3299161+celynw@users.noreply.github.com" },
]
requires-python = ">=3.14"
dependencies = [
	"kellog>=1.0.1",
	"natsort>=8.4.0",
	"platformdirs>=4.5.0",
	"pyyaml>=6.0.3",
	"steamfront>=0.1.0",
	"typer>=0.24",
]

[project.scripts]
steam_screenshot_manager = "steam_screenshot_manager.main:app"

[dependency-groups]
dev = [
	"mypy>=1.0",
	"pytest>=8.0",
	"pytest-cov>=5.0",
	"ruff>=0.9",
	"types-PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/steam_screenshot_manager"]

[tool.hatch.build]
include = ["src/steam_screenshot_manager/replacements.yml"]

# Ruff
[tool.ruff]
target-version = "py314"
line-length = 88

[tool.ruff.lint]
select = ["ALL"]
ignore = [
	"ANN003", # Missing type annotation for **kwargs
	"CPY001", # Missing copyright notice at top of file (LICENSE file is sufficient)
	"D107",   # Missing docstring in `__init__` (put it on the class instead)
	"D203",   # 1 blank line required before class docstring (conflicts with D211)
	"D206",   # Docstring should be indented with spaces, not tabs (formatter handles this)
	"D212",   # Multi-line docstring summary should start at first line (conflicts with D213)
	"D413",   # Missing blank line after last section
	"E203",   # Whitespace before ':' (conflicts with black slice style `list[1 : 1 + 1]`)
	"ERA001", # Found commented-out code
	"ISC001", # Implicitly concatenated string literals (conflicts with ruff formatter)
	"TD003",  # Missing issue link on the line following this TODO
	"W191",   # Indentation contains tabs (formatter handles this)
	"COM812", # Missing trailing comma in a multi-line collection (formatter handles this)
]
unfixable = [
	"B007", # Loop control variable not used in loop body
	"B905", # `zip()` without an explicit `strict=` parameter
	"D204", # 1 blank line required after class docstring
	"F401", # Imported but unused
	"F841", # Local variable is assigned to but never used
]

[tool.ruff.lint.extend-per-file-ignores]
"**/__init__.py" = [
	"D104", # Missing docstring in public package
]
"test/**/*.py" = [
	"ANN001",  # Allow untyped pytest fixtures like `monkeypatch`
	"ARG005",  # Allow intentionally unused lambda args in monkeypatched stubs
	"D100",    # Missing docstring in public module
	"D103",    # Missing docstring in public function
	"DOC201",  # `return` is not documented in docstring
	"FBT",     # Boolean positional args (common in @pytest.mark.parametrize)
	"INP001",  # File is part of an implicit namespace package
	"PLR2004", # Magic value used in comparison
	"S108",    # Allow stable test paths like /tmp fixtures in test literals
	"SLF001",  # Private member access is acceptable in tests
	"S101",    # Use of `assert` detected
	"S311",    # Standard pseudo-random generators
	"TC003",   # Keep test imports straightforward instead of TYPE_CHECKING blocks
]

[tool.ruff.lint.isort]
no-lines-before = ["local-folder"]
split-on-trailing-comma = true
combine-as-imports = true
force-wrap-aliases = true

[tool.ruff.lint.flake8-annotations]
mypy-init-return = true

[tool.ruff.lint.flake8-unused-arguments]
ignore-variadic-names = true

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["typer.Argument", "typer.Option"]

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.format]
indent-style = "tab"
quote-style = "double"
docstring-code-format = true
line-ending = "lf"
skip-magic-trailing-comma = false

# Mypy
[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true

# Add per-library overrides here as needed, e.g.:
# [[tool.mypy.overrides]]
# module = "some_untyped_lib.*"
# ignore_missing_imports = true

# Pytest
[tool.pytest.ini_options]
testpaths = ["test"]
pythonpath = ["src"]
addopts = [
	"--cov=steam_screenshot_manager",
	"--cov-report=term-missing",
	"--cov-report=html",
]
